backend/requirements.txt
fastapi>=0.115.0
slowapi>=0.1.9
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.6.1
alembic>=1.14.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
email-validator>=2.2.0
reportlab>=4.0.0
httpx>=0.27.0
